pyproject.toml
[tool.poetry]
name = "upr-k-backend"
version = "0.1.0"
description = "Backend of UPR-K"
authors = [
    "Rafael Martinez Estevez <malayibiri@gmail.com>",
    "Sergio Perez Reyes <sergioreperez922@gmail.com>",
    "LázaroHernándezPérez<lhdezp0211@gmail.com>",
    "Javier Valladares Alonso<valladaresalonsojavieres@gmail.com>"
]
license = "GNU"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.1"
sqlparse = "0.4.4"
tzdata = "2023.3"
asgiref = "3.7.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
